# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yet Another Node.js Blog Application API",
    "description": "Yet Another Node.js Blog Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/review/{resId}/{reviewId}": {
      "parameters": [
        {
          "name": "resId",
          "in": "path",
          "required": true,
          "description": "ID of the restaurant",
          "type": "string"
        },
        {
          "name": "reviewId",
          "in": "path",
          "required": true,
          "description": "ID os the review",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Review"
        ],
        "description": "List of all reviews",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number, Default to Zero"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of reviews with page size of 50",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/Reviews"
                }
              }
            }
          },
          "4xx": {
            "description": "Fetching failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Review"
        ],
        "description": "Create new review in system",
        "parameters": [
          {
            "name": "review",
            "in": "body",
            "description": "Review that we want to create",
            "schema": {
              "properties": {
                "rating": {
                  "type": "string",
                  "example": "rating"
                },
                "comment": {
                  "type": "string",
                  "example": "comment"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The review is deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/definitions/Reviews"
                    }
                  }
                }
              }
            }
          },
          "40x": {
            "description": "When review updation is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Review"
        ],
        "description": "Create new review in system",
        "parameters": [
          {
            "name": "review",
            "in": "body",
            "description": "Review that we want to create",
            "schema": {
              "properties": {
                "reply": {
                  "type": "string",
                  "example": "reply"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The review is deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/definitions/Reviews"
                    }
                  }
                }
              }
            }
          },
          "40x": {
            "description": "When review creation is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Review"
        ],
        "description": "Delete review in system",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The review is deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/definitions/Reviews"
                    }
                  }
                }
              }
            }
          },
          "40x": {
            "description": "When review deletion is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/review/{resId}/{reviewId}/reply": {
      "parameters": [
        {
          "name": "resId",
          "in": "path",
          "required": true,
          "description": "ID of the restaurant",
          "type": "string"
        },
        {
          "name": "reviewId",
          "in": "path",
          "required": true,
          "description": "ID os the review",
          "type": "string"
        }
      ],
      "put": {
        "tags": [
          "Review"
        ],
        "description": "Create new reply in system",
        "parameters": [
          {
            "name": "review",
            "in": "body",
            "description": "Reply that we want to create",
            "schema": {
              "properties": {
                "reply": {
                  "type": "string",
                  "example": "reply"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The reply is deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "properties": {
                        "reply": {
                          "type": "string",
                          "example": "reply"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "40x": {
            "description": "When reply updation is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Review"
        ],
        "description": "Delete reply in system",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "The reply is deleted",
            "schema": {
              "properties": {
                "success": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "properties": {
                        "reply": {
                          "type": "string",
                          "example": "reply"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "40x": {
            "description": "When reply deletion is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/review/{resId}": {
      "parameters": [
        {
          "name": "resid",
          "in": "path",
          "required": true,
          "description": "ID of the restaurant",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Review"
        ],
        "description": "List of all reviews",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number, Default to Zero"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List of reviews with page size of 50",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Reviews"
                  }
                }
              }
            }
          },
          "4xx": {
            "description": "Fetching failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Review"
        ],
        "description": "Create new review in system",
        "parameters": [
          {
            "name": "review",
            "in": "body",
            "description": "Review that we want to create",
            "schema": {
              "properties": {
                "comment": {
                  "type": "string",
                  "example": "comment"
                },
                "rating": {
                  "type": "float",
                  "example": 4.5
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "New review is created",
            "schema": {
              "$ref": "#/definitions/SuccessUserResponse"
            }
          },
          "40x": {
            "description": "When review creation is failed",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/review": {
      "get": {
        "tags": [
          "Review"
        ],
        "description": "List of reviews applicable for users",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page number, Default to Zero",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Reviews"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Reviews": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "rating": {
          "type": "float",
          "example": 4.5
        },
        "comment": {
          "type": "string",
          "example": "comment"
        },
        "reply": {
          "type": "string",
          "example": "reply"
        },
        "dateOfVisit": {
          "type": "string",
          "example": "ISO Format date"
        },
        "createdAt": {
          "type": "string",
          "example": "ISO Format date"
        },
        "updatedAt": {
          "type": "string",
          "example": "ISO Format date"
        }
      }
    },
...
    "SuccessUserResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "description": "Success message"
            }
          }
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}